backend/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
